requirements.txt
# Core dependencies
openai>=1.0.0
google-api-python-client>=2.100.0
oauth2client>=4.1.3
openai>=1.0.0
google-api-python-client>=2.100.0
oauth2client>=4.1.3
reportlab>=4.0.0
python-docx>=0.8.11
pandas>=2.0.0
pydantic
pydantic-settings

# Google Sheets integration
gspread>=6.0.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Support libraries
httplib2>=0.20.4
argparse>=1.4.0
python-dateutil>=2.8.2

# UI dependencies
streamlit==1.36.0
markdown2>=2.4.8

# Development dependencies
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
python-docx>=0.8.11
pandas>=2.0.0

# Support libraries
httplib2>=0.20.4
argparse>=1.4.0
python-dateutil>=2.8.2

# UI dependencies
streamlit==1.36.0
markdown2>=2.4.8